WIP: computing results
The aot calibration estimate is causing problems. Probably too
much variability from the start to the finish, particularly in
the very slow-running debug mode we're currently using (~1.3 s
between packets used for the calibration).

Not sure if it make sense to start pushing timing a bit now to
give this a fighting chance or rework to the protocol.