Dial down printf speed
This is max speed when printing all 27 range measurements (barring
some other tricks to reduce the number of calls to printf, but I
don't think those will buy us so much more). It's ~6.67 Hz.

The range event takes ~33 ms. printf'ing the results takes ~125 ms.

The next step is 10%ile calculations on the tag to reduce the volume
that is printed out.